requirements.txt
pytest>=8.2.2
requests>=2.25.1
pytest-html>=3.1.1
unittest-xml-reporting>=3.0.4
pytest-cov>=4.0.0

pytest-playwright